Extract ProductSummaryList without query
